MapItemInputHeader Allow Edit of UomName and allow to add null price

+++ Features/Admin/PriceIncrease/DTOs/PriceIncreaseDtos.cs
@@ -52,7 +52,7 @@ namespace STTproject.Features.Admin.PriceIncrease.DTOs
         public string? SubdItemName { get; set; }
         public int ItemsUomId { get; set; }
         public string? UomName { get; set; }
-        public decimal? ConversionToBase { get; set; }
+        public int? ConversionToBase { get; set; }
         public decimal? OldPrice { get; set; }
         public decimal? NewPrice { get; set; }
